added seperate page for habits
