+++ Habit Tracker/package.json
@@ -9,6 +9,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.3.0",
+    "react-router-dom": "^7.1.3",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
