+++ Habit Tracker/src/styling/HabitListPage.css
@@ -1,0 +1,68 @@
+/* Ensure .habit-page respects the parent container dimensions */
+.habit-page {
+    max-width: 1200px; /* Adjust the width as needed */
+    width: 100%;
+    margin: 0 auto;
+    padding: 30px; /* Add breathing room */
+    background: #ecf1f2; /* Keep it consistent with the app's style */
+    border-radius: 20px;
+    box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2); /* Enhanced shadow for prominence */
+}
+
+.habit-page-title {
+    font-size: 24px;
+    font-weight: bold;
+    color: #333;
+    text-align: center;
+    margin-bottom: 20px;
+}
+/* Habit list item improvements */
+.habit-list {
+    display: flex;
+    flex-direction: column;
+    gap: 20px; /* Space between each card */
+}
+
+.habit-item {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 20px;
+    background: linear-gradient(120deg, #bb7070, #92322d); /* Subtle gradient */
+    border-radius: 15px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1); /* Subtle lift effect */
+    transition: transform 0.3s, background-color 0.3s, box-shadow 0.3s;
+}
+
+.habit-item:hover {
+    transform: translateY(-5px); /* Lift on hover */
+    background: linear-gradient(120deg, #bb7070, #92322d);
+    box-shadow: 0 12px 24px rgba(0, 0, 0, 0.15);
+}
+
+/* Override completed habit style */
+.habit-item.completed {
+    text-decoration: line-through;
+    color: #8c8c8c;
+    background:linear-gradient(120deg, #5be6aa, #50a395); /* Muted background for completed habits */
+    opacity: 0.9;
+}
+
+.habit-item span {
+    font-size: 1.2rem;
+    font-weight: 500;
+    color: #333;
+}
+
+/* Responsive adjustments */
+@media (max-width: 768px) {
+    .habit-page {
+        padding: 20px;
+        padding-left: 10%;
+    }
+
+    .habit-item {
+        flex-direction: column; /* Stack elements on smaller screens */
+        align-items: flex-start;
+    }
+}
